# Auto vectorize `replace_copy`, `replace_copy_if`