--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3851,8 +3851,8 @@ _CONSTEXPR20 _OutIt replace_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const 
     const auto _ULast = _STD _Get_unwrapped(_Last);
     auto _UDest       = _STD _Get_unwrapped_n(_Dest, _STD _Idl_distance<_InIt>(_UFirst, _ULast));
     for (; _UFirst != _ULast; ++_UFirst, (void) ++_UDest) {
-        if constexpr (is_integral_v<_Ty> && is_integral_v<iter_value_t<_InIt>>
-                      || is_pointer_v<_Ty> && is_pointer_v<iter_value_t<_InIt>>) {
+        if constexpr ((is_integral_v<_Ty> && is_integral_v<_Iter_value_t<_InIt>>)
+                      || (is_pointer_v<_Ty> && is_pointer_v<_Iter_value_t<_InIt>>) ) {
             // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
             // by making transformation that can be done for simple types
             *_UDest = (*_UFirst == _Oldval) ? _Newval : *_UFirst;
@@ -3928,8 +3928,8 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<equal_to, projected<_It, _Pj>, const _Ty1*>);
 
             for (; _First != _Last; ++_First, (void) ++_Result) {
-                if constexpr (is_integral_v<_Ty2> && is_integral_v<iter_value_t<_It>>
-                              || is_pointer_v<_Ty2> && is_pointer_v<iter_value_t<_It>>) {
+                if constexpr ((is_integral_v<_Ty2> && is_integral_v<iter_value_t<_It>>)
+                              || (is_pointer_v<_Ty2> && is_pointer_v<iter_value_t<_It>>) ) {
                     // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
                     // by making transformation that can be done for simple types
                     *_Result = _STD invoke(_Proj, *_First) == _Oldval ? _Newval : *_First;