--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3843,6 +3843,8 @@ namespace ranges {
 } // namespace ranges
 #endif // _HAS_CXX20
 
+// TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
+// by making transformation that can be done for simple types
 template <class _UOutIt, class _InTy, class _NewTy>
 constexpr bool _Can_vectorize_replace =
     conjunction_v<is_pointer<_UOutIt>, is_same<_InTy, _NewTy>, disjunction<is_integral<_InTy>, is_pointer<_InTy>>>;
@@ -3856,8 +3858,6 @@ _CONSTEXPR20 _OutIt replace_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const 
     auto _UDest       = _STD _Get_unwrapped_n(_Dest, _STD _Idl_distance<_InIt>(_UFirst, _ULast));
     for (; _UFirst != _ULast; ++_UFirst, (void) ++_UDest) {
         if constexpr (_Can_vectorize_replace<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
-            // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
-            // by making transformation that can be done for simple types
             *_UDest = (*_UFirst == _Oldval) ? _Newval : *_UFirst;
         } else {
             if (*_UFirst == _Oldval) {
@@ -3931,9 +3931,7 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<equal_to, projected<_It, _Pj>, const _Ty1*>);
 
             for (; _First != _Last; ++_First, (void) ++_Result) {
-                if constexpr (_Can_vectorize_replace<_Out, _Iter_value_t<_It>, _Ty2>) {
-                    // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
-                    // by making transformation that can be done for simple types
+                if constexpr (_Can_vectorize_replace<_Out, iter_value_t<_It>, _Ty2>) {
                     *_Result = _STD invoke(_Proj, *_First) == _Oldval ? _Newval : *_First;
                 } else {
                     if (_STD invoke(_Proj, *_First) == _Oldval) {
@@ -3960,10 +3958,14 @@ _CONSTEXPR20 _OutIt replace_copy_if(_InIt _First, _InIt _Last, _OutIt _Dest, _Pr
     const auto _ULast = _STD _Get_unwrapped(_Last);
     auto _UDest       = _STD _Get_unwrapped_n(_Dest, _STD _Idl_distance<_InIt>(_UFirst, _ULast));
     for (; _UFirst != _ULast; ++_UFirst, (void) ++_UDest) {
-        if (_Pred(*_UFirst)) {
-            *_UDest = _Val;
+        if constexpr (_Can_vectorize_replace<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
+            *_UDest = _Pred(*_UFirst) ? _Val : *_UFirst;
         } else {
-            *_UDest = *_UFirst;
+            if (_Pred(*_UFirst)) {
+                *_UDest = _Val;
+            } else {
+                *_UDest = *_UFirst;
+            }
         }
     }
 
@@ -4030,10 +4032,14 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_unary_predicate<_Pr, projected<_It, _Pj>>);
 
             for (; _First != _Last; ++_First, (void) ++_Result) {
-                if (_STD invoke(_Pred, _STD invoke(_Proj, *_First))) {
-                    *_Result = _Newval;
+                if constexpr (_Can_vectorize_replace<_Out, iter_value_t<_It>, _Ty>) {
+                    *_Result = _STD invoke(_Pred, _STD invoke(_Proj, *_First)) ? _Newval : *_First;
                 } else {
-                    *_Result = *_First;
+                    if (_STD invoke(_Pred, _STD invoke(_Proj, *_First))) {
+                        *_Result = _Newval;
+                    } else {
+                        *_Result = *_First;
+                    }
                 }
             }