--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3846,7 +3846,7 @@ namespace ranges {
 // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
 // by making transformation that can be done for simple types
 template <class _UOutIt, class _InTy, class _NewTy>
-constexpr bool _Can_vectorize_replace =
+constexpr bool _Can_vectorize_replace_copy =
     conjunction_v<is_pointer<_UOutIt>, is_same<_InTy, _NewTy>, disjunction<is_integral<_InTy>, is_pointer<_InTy>>>;
 
 _EXPORT_STD template <class _InIt, class _OutIt, class _Ty>
@@ -3857,7 +3857,7 @@ _CONSTEXPR20 _OutIt replace_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const 
     const auto _ULast = _STD _Get_unwrapped(_Last);
     auto _UDest       = _STD _Get_unwrapped_n(_Dest, _STD _Idl_distance<_InIt>(_UFirst, _ULast));
     for (; _UFirst != _ULast; ++_UFirst, (void) ++_UDest) {
-        if constexpr (_Can_vectorize_replace<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
+        if constexpr (_Can_vectorize_replace_copy<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
             *_UDest = (*_UFirst == _Oldval) ? _Newval : *_UFirst;
         } else {
             if (*_UFirst == _Oldval) {
@@ -3931,7 +3931,7 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<equal_to, projected<_It, _Pj>, const _Ty1*>);
 
             for (; _First != _Last; ++_First, (void) ++_Result) {
-                if constexpr (_Can_vectorize_replace<_Out, iter_value_t<_It>, _Ty2>) {
+                if constexpr (_Can_vectorize_replace_copy<_Out, iter_value_t<_It>, _Ty2>) {
                     *_Result = _STD invoke(_Proj, *_First) == _Oldval ? _Newval : *_First;
                 } else {
                     if (_STD invoke(_Proj, *_First) == _Oldval) {
@@ -3958,7 +3958,7 @@ _CONSTEXPR20 _OutIt replace_copy_if(_InIt _First, _InIt _Last, _OutIt _Dest, _Pr
     const auto _ULast = _STD _Get_unwrapped(_Last);
     auto _UDest       = _STD _Get_unwrapped_n(_Dest, _STD _Idl_distance<_InIt>(_UFirst, _ULast));
     for (; _UFirst != _ULast; ++_UFirst, (void) ++_UDest) {
-        if constexpr (_Can_vectorize_replace<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
+        if constexpr (_Can_vectorize_replace_copy<decltype(_UDest), _Iter_value_t<_InIt>, _Ty>) {
             *_UDest = _Pred(*_UFirst) ? _Val : *_UFirst;
         } else {
             if (_Pred(*_UFirst)) {
@@ -4032,7 +4032,7 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_unary_predicate<_Pr, projected<_It, _Pj>>);
 
             for (; _First != _Last; ++_First, (void) ++_Result) {
-                if constexpr (_Can_vectorize_replace<_Out, iter_value_t<_It>, _Ty>) {
+                if constexpr (_Can_vectorize_replace_copy<_Out, iter_value_t<_It>, _Ty>) {
                     *_Result = _STD invoke(_Pred, _STD invoke(_Proj, *_First)) ? _Newval : *_First;
                 } else {
                     if (_STD invoke(_Pred, _STD invoke(_Proj, *_First))) {