--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -3846,8 +3846,12 @@ namespace ranges {
 // TRANSITION, DevCom-10606350: help the compiler auto-vectorizing for simple types
 // by making transformation that can be done for simple types
 template <class _UOutIt, class _InTy, class _NewTy>
-constexpr bool _Can_vectorize_replace_copy =
-    conjunction_v<is_pointer<_UOutIt>, is_same<_InTy, _NewTy>, disjunction<is_integral<_InTy>, is_pointer<_InTy>>>;
+constexpr bool _Can_vectorize_replace_copy = conjunction_v<is_pointer<_UOutIt>, is_same<_InTy, _NewTy>,
+    disjunction<
+#ifdef __cpp_lib_byte
+        is_same<_InTy, byte>,
+#endif // defined(__cpp_lib_byte)
+        is_integral<_InTy>, is_pointer<_InTy>>>;
 
 _EXPORT_STD template <class _InIt, class _OutIt, class _Ty>
 _CONSTEXPR20 _OutIt replace_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const _Ty& _Oldval, const _Ty& _Newval) {