@AlexGuteniev 2024-03-08
> I don't think we need to require the output type to be the same as the source types.

Curiously, the AVX2 auto vectorization does indeed work even for different source and destination element size, this is achieved using `vpmovsx*` or `vpmovzx*` to widen and `vpshufb` to compress.

Surprising the compiles goes that much far, and misses the final piece to vectorize without this PR