@AlexGuteniev 2024-03-20
Noticed a bug here, will push a change after CI passes.
I didn't realize initially that projection can change type, so may break the ternary.
We should check it for being `identity`.