@StephanTLavavej 2024-03-20
Good idea to think about the projection - I hadn't explicitly done so. However, does it matter? The condition is being moved monolithically from the `if`-statement to the ternary operator, so my understanding is that the projection (and predicate) can be arbitrarily squirrelly without damaging the semantics of this transformation.