@AlexGuteniev 2024-03-20
> However, does it matter? The condition is being moved monolithically from the `if`-statement to the ternary operator, so my understanding is that the projection (and predicate) can be arbitrarily squirrelly without damaging the semantics of this transformation

Yes, you are right. It doesn't matter what type is in the conditional of ternary, only what types are in resulting expressions.

Initially I event tested that simple projections still vectorize along with the whole algorithm.