@AlexGuteniev 2025-04-26
This no longer works in latest Preview. 
Not sure when exactly it was broken, but works in latest Production.
This is compiler change, not STL change.

DevCom-10895463